+++ components/surfing.tsx
@@ -1,0 +1,34 @@
+interface SurfingProps {
+  className?: string;
+}
+
+export function Surfing({ className }: SurfingProps) {
+  return (
+    <pre className={className}>
+      <code>{`
+                       ..
+                   .+....:
+                   :~.~+=*
+                  .:....~+       ..
+              +++++**..=}}[())<)(:
+          ~^)(][[[[[[}{{{)}}}]^
+        -(]}[)-){#{{{{{{~
+       <[{(   -)#######(
+      ..^     *[{{{{#{}(<>^.
+     ..       ^(](]}#{}[(<><)
+              ++<><(   :){{{{.   .~:
+              -=^*>)     .##{=.--   .~::
+               *^><)  :===}#{-    ~::+*
+              +^]][]===~..:*^-..-.-++
+            -^({{{>=~ --    := :=++
+           =^}}<=.:-.    :-.:~=++
+         =~=-- -~     :~.:===++
+      :++~ -...   .~- -====*~
+    .~-:~-  :::-~-:~===++=
+   -+~     :=~ ~====+*~
+  .    :~---=++**+:
+  ~+++=++**+:
+        `}</code>
+    </pre>
+  );
+}
