+++ package.json
@@ -31,10 +31,9 @@
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "e2b": "^1.0.7",
-    "framer-motion": "^11.15.0",
     "hls.js": "^1.5.19",
     "lucide-react": "^0.474.0",
-    "motion": "^12.4.3",
+    "motion": "^12.5.0",
     "next": "^15.3.0-canary.0",
     "next-themes": "^0.4.4",
     "openai": "4.87.2",
